requirements.txt
```txt
streamlit
transformers
torch
gtts
Pillow

streamlit==1.34.0
transformers==4.41.1
torch>=2.0.0
gtts==2.5.1
Pillow==9.5.0
